fix(quadraui): isolate pty smoke tests from ambient COLORTERM, correct SGR docs (#826)
Review fix iteration 1:

- tests/tui_pty_smoke.rs: PtyExample::spawn_with_env now env_remove's
  COLORTERM before applying the caller's extra_env. portable-pty's
  CommandBuilder seeds its env map from the entire parent process
  environment, so an ambient COLORTERM=truecolor in the shell running
  `cargo test` (VS Code's integrated terminal, tmux with passthrough,
  many terminal emulators) was silently leaking into the spawned
  example and could make detect_color_depth() resolve to TrueColor
  regardless of the test's intended TERM — making the two
  acceptance-critical fixtures (indexed-under-256color,
  ansi16-under-plain-xterm) host-dependent rather than deterministic.
- src/tui/color.rs, src/backend.rs: corrected doc comments that
  claimed the Ansi16 path emits classic 30-37/90-97 SGR codes.
  crossterm 0.29's Colored Display impl formats every named Color
  variant via the extended 38;5;n/48;5;n form, matching the PR's own
  ANSI16_PIPELINE_SGR fixture constant — Ansi16 narrows the palette,
  not the escape-sequence family. Not a behavior change, doc-only.
- src/tui/backend.rs: import ColorDepth instead of fully qualifying
  crate::backend::ColorDepth at each of the three use sites.
- src/tui/color.rs: dist2 uses i32 instead of i64 — squared RGB
  distance never exceeds 255*255*3, i32 is ample and matches the
  values' actual range.

Verified: `cargo test --features tui,terminal --test tui_pty_smoke`
now passes deterministically; full tui-leg quality gate (build, lib
tests, clippy -D warnings, fmt --check) is clean.

+++ quadraui/src/tui/color.rs
@@ -65,10 +65,10 @@ const CUBE_LEVELS: [u8; 6] = [0, 95, 135, 175, 215, 255];
 /// Squared Euclidean distance between two RGB triples — cheap and
 /// monotonic with true distance, which is all nearest-colour picking
 /// needs.
-fn dist2(a: (u8, u8, u8), b: (u8, u8, u8)) -> i64 {
-    let dr = a.0 as i64 - b.0 as i64;
-    let dg = a.1 as i64 - b.1 as i64;
-    let db = a.2 as i64 - b.2 as i64;
+fn dist2(a: (u8, u8, u8), b: (u8, u8, u8)) -> i32 {
+    let dr = a.0 as i32 - b.0 as i32;
+    let dg = a.1 as i32 - b.1 as i32;
+    let db = a.2 as i32 - b.2 as i32;
     dr * dr + dg * dg + db * db
 }
 
@@ -154,9 +154,17 @@ const ANSI16_PALETTE: [(RatatuiColor, (u8, u8, u8)); 16] = [
     (RatatuiColor::White, (255, 255, 255)),
 ];
 
-/// Nearest of the 16 basic ANSI colours for `(r, g, b)`, as the
-/// [`RatatuiColor`] named variant crossterm serialises to plain `3x`/`9x`
-/// SGR codes (never `38;5;n`).
+/// Nearest of the 16 basic ANSI colours for `(r, g, b)`.
+///
+/// This narrows the *palette* to the 16 basic ANSI colours, but does not
+/// change the *escape-sequence family*: crossterm 0.29's `Colored` `Display`
+/// impl formats every named [`RatatuiColor`] variant — including
+/// `Black..=White` — via the extended 8-bit form (`38;5;n` / `48;5;n`), the
+/// same family [`rgb_to_indexed256`] produces; there is no crossterm API
+/// path that emits classic `30-37`/`90-97` SGR codes. So this fallback
+/// helps a terminal that supports extended-256 syntax but only renders/
+/// themes 16 colours; it does not by itself help a terminal that only
+/// understands classic 3/4-bit SGR (a real vt100, a bare serial console).
 pub(crate) fn rgb_to_ansi16(r: u8, g: u8, b: u8) -> RatatuiColor {
     ANSI16_PALETTE
         .iter()
